Refactor: domain package 생성
