+++ src/main/java/com/example/mangoplace/domain/review/dto/request/UpdateReviewRequest.java
@@ -1,0 +1,15 @@
+package com.example.mangoplace.domain.review.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class UpdateReviewRequest {
+    private Integer star;
+    private String content;
+}
