+++ src/main/java/com/example/mangoplace/domain/signup/controller/TestController.java
@@ -1,0 +1,36 @@
+package com.example.mangoplace.domain.signup.controller;
+
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin(origins = "*", maxAge = 3600)
+@RestController
+@RequestMapping("/api/test")
+public class TestController {
+    @GetMapping("/all")
+    public String allAccess() {
+        return "공용 콘텐츠";
+    }
+
+    @GetMapping("/user")
+    //권한이 있는지 검사
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public String userAccess() {
+        return "유저 콘텐츠";
+    }
+//
+//    @GetMapping("/ed")
+//    @PreAuthorize("hasRole('EDITOR')")
+//    public String moderatorAccess() {
+//        return "Editor Board.";
+//    }
+//
+//    @GetMapping("/admin")
+//    @PreAuthorize("hasRole('ADMIN')")
+//    public String adminAccess() {
+//        return "Admin Board.";
+//    }
+}
