+++ src/main/java/com/example/mangoplace/domain/signup/entity/ERole.java
@@ -1,0 +1,6 @@
+package com.example.mangoplace.domain.signup.entity;
+
+public enum ERole {
+    ROLE_USER,
+    ROLE_ADMIN //admin = 관리자
+}
