+++ src/main/java/com/example/mangoplace/domain/signup/payload/response/MessageResponse.java
@@ -1,0 +1,17 @@
+package com.example.mangoplace.domain.signup.payload.response;
+
+public class MessageResponse {
+    private String message;
+
+    public MessageResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
