+++ src/main/java/com/example/mangoplace/domain/signup/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package com.example.mangoplace.domain.signup.repository;
+
+import com.example.mangoplace.domain.signup.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+
+    Boolean existsByUsername(String username);
+
+    Boolean existsByEmail(String email);
+}
