+++ src/main/java/com/example/mangoplace/domain/signup/service/UserDetailsServiceImpl.java
@@ -1,0 +1,29 @@
+package com.example.mangoplace.domain.signup.service;
+
+import com.example.mangoplace.domain.signup.repository.UserRepository;
+import com.example.mangoplace.domain.signup.entity.User;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class UserDetailsServiceImpl implements UserDetailsService {
+    private UserRepository userRepository;
+
+    public UserDetailsServiceImpl(UserRepository userRepository){
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    @Transactional
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User Not Found with username: " + username));
+
+        return UserDetailsImpl.build(user);
+    }
+
+}
