--- src/main/java/com/example/mangoplace/review/dto/response/UpdateReviewResponse.java
@@ -1,33 +1,0 @@
-package com.example.mangoplace.review.dto.response;
-
-import com.example.mangoplace.review.dto.request.UpdateReviewRequest;
-import com.example.mangoplace.review.entity.ReviewEntity;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UpdateReviewResponse {
-    private Long reviewId;
-    private LocalDateTime updatedAt;
-    private Long userId;
-    private String userName;
-    private String content;
-    private int star;
-
-    public static UpdateReviewResponse fromEntity(ReviewEntity review){
-        return UpdateReviewResponse.builder()
-                .reviewId(review.getId())
-                .updatedAt(review.getUpdatedAt())
-                .star(review.getStar())
-                .userId(review.getUser().getId())
-                .userName(review.getUser().getUsername())
-                .build();
-    }
-}
