--- src/main/java/com/example/mangoplace/review/entity/ReviewEntity.java
@@ -1,48 +1,0 @@
-package com.example.mangoplace.review.entity;
-
-import com.example.mangoplace.review.dto.request.UpdateReviewRequest;
-import com.example.mangoplace.signup.entity.User;
-import jakarta.persistence.*;
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-@Builder
-public class ReviewEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String content;
-
-    private Integer star;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @CreatedDate
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    public void update(UpdateReviewRequest reviewUpdateRequest) {
-        this.star = reviewUpdateRequest.getStar();
-        this.content = reviewUpdateRequest.getContent();
-    }
-
-}
-
