--- src/main/java/com/example/mangoplace/signup/entity/ERole.java
@@ -1,6 +1,0 @@
-package com.example.mangoplace.signup.entity;
-
-public enum ERole {
-    ROLE_USER,
-    ROLE_ADMIN //admin = 관리자
-}
