--- src/main/java/com/example/mangoplace/signup/entity/User.java
@@ -1,82 +1,0 @@
-package com.example.mangoplace.signup.entity;
-
-import com.example.mangoplace.review.entity.ReviewEntity;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-@Table(name = "users")
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    //유저네임
-    @Column(length = 20, nullable = false, unique = true)
-    private String username;
-
-    @Column(length = 120, nullable = false)
-    private String password;
-
-//    //닉네임
-//    @Column(length = 20, nullable = false, unique = true)
-//    private String nickname;
-
-    @Email
-    @Column(length = 50, nullable = false, unique = true)
-    private String email;
-
-    @ManyToMany(fetch = FetchType.LAZY)
-    @JoinTable(name = "user_roles",
-            joinColumns = @JoinColumn(name = "user_id"),
-            inverseJoinColumns = @JoinColumn(name = "role_id"))
-    private Set<Role> roles = new HashSet<>();
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private List<ReviewEntity> review = new ArrayList<>();
-
-    @Builder
-    public User(String username, String email, String password) {
-        this.username = username;
-        this.email = email;
-        this.password = password;
-    }
-
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-
-    public void setRoles(Set<Role> roles) {
-        this.roles = roles;
-    }
-}
