--- src/main/java/com/example/mangoplace/signup/payload/request/LoginRequest.java
@@ -1,27 +1,0 @@
-package com.example.mangoplace.signup.payload.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public class LoginRequest {
-    @NotBlank
-    private String username;
-
-    @NotBlank
-    private String password;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
