--- src/main/java/com/example/mangoplace/signup/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.example.mangoplace.signup.repository;
-
-import com.example.mangoplace.signup.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-
-    Boolean existsByUsername(String username);
-
-    Boolean existsByEmail(String email);
-}
