--- src/main/java/com/example/mangoplace/signup/service/UserDetailsServiceImpl.java
@@ -1,30 +1,0 @@
-package com.example.mangoplace.signup.service;
-
-import com.example.mangoplace.signup.entity.User;
-import com.example.mangoplace.signup.repository.UserRepository;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-    private UserRepository userRepository;
-
-    public UserDetailsServiceImpl(UserRepository userRepository){
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User Not Found with username: " + username));
-
-        return UserDetailsImpl.build(user);
-    }
-
-}
